% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbMetaClient-methods.R
\docType{methods}
\name{cbMetaClient,CellBaseR-method}
\alias{cbMetaClient}
\alias{cbMetaClient,CellBaseR-method}
\title{A method for getting the avaiable metadata from the cellbase web services}
\usage{
\S4method{cbMetaClient}{CellBaseR}(object, resource)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{resource}{the resource you want to query it metadata}
}
\value{
a dataframe with the
 results of the query
}
\description{
This method is for getting species data from the cellbase web services.
}
\details{
Please, for details on possible values for the 
parameters  and  additional filters of this function refer to 
https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbMetaClient(object=cb, resource="species")
}

