% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createGeneModel}
\alias{createGeneModel}
\title{A convience functon to construct a genemodel}
\usage{
createGeneModel(object, region = NULL)
}
\arguments{
\item{object}{an object of class CellbaseResponse}

\item{region}{a character}
}
\value{
A geneModel
}
\description{
A convience functon to construct a genemodel
}
\details{
This function takes cbResponse object and returns a geneModel, 
this geneModel can be then turned into a GeneRegion Track object by Gviz
}
\examples{
cb <- CellBaseR()
test <- createGeneModel(object = cb, region = "17:1500000-1550000")
}

