% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{getCellBaseResourceHelp}
\alias{getCellBaseResourceHelp}
\title{A function to get help about available cellbase resources}
\usage{
getCellBaseResourceHelp(object, subcategory)
}
\arguments{
\item{object}{a cellBase class object}

\item{subcategory}{a character the subcategory to be queried}
}
\value{
documentation about available resources rs
}
\description{
This is a convience function to get help on cellbase methods
}
\examples{
cb <- CellBaseR()
getCellBaseResourceHelp(cb, subcategory="gene")
}
