% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromosomeInfo-methods.R
\docType{methods}
\name{getChromosomeInfo,CellBaseR-method}
\alias{getChromosomeInfo,CellBaseR-method}
\alias{getChromosomeInfo}
\title{getChromosomeInfo}
\usage{
\S4method{getChromosomeInfo}{CellBaseR}(object, ids, resource, param = NULL,
  ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of chromosome ids to be queried}

\item{resource}{a character vector to specify the resource to be queried}

\item{param}{a object of class CellBaseParam specifying additional param for the query}

\item{...}{any extra arguments}
}
\value{
a dataframe with the results of the query
}
\description{
A method to query sequence data from Cellbase web services.
}
\details{
Please, for details on possible values for the 
parameters  and  additional param of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- getChromosomeInfo(object=cb, ids="22", resource="info")
}
