% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getClinicalByGene}
\alias{getClinicalByGene}
\title{getClinicalByGene}
\usage{
getClinicalByGene(gene)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch clinical variants for specific gene/s
}
\examples{
res <- getClinicalByGene("TET1")
}

