% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getClinicalByRegion}
\alias{getClinicalByRegion}
\title{getClinicalByRegion}
\usage{
getClinicalByRegion(region, object = NULL, filters = NULL)
}
\arguments{
\item{region}{a charcter vector of genomic regions, eg 17:1000000-1100000}

\item{object}{an object of class CellBaseR}

\item{filters}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch clinical variants for specific region/s
}
\examples{
res <- getClinicalByRegion("17:1000000-1189811")
}
