% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{getGeneInfo}
\usage{
getGeneInfo(gene)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch gene annotations specific gene/s
}
\examples{
res <- getGeneInfo("TET1")
}

