% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta-methods.R
\docType{methods}
\name{getMeta,CellBaseR-method}
\alias{getMeta}
\alias{getMeta,CellBaseR-method}
\title{getMeta}
\usage{
\S4method{getMeta}{CellBaseR}(object, resource)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{resource}{the resource you want to query it metadata}
}
\value{
a dataframe with the
 results of the query
}
\description{
A method for getting the avaiable metadata from the cellbase web services
}
\details{
This method is for getting species data from the cellbase web services.

Please, for details on possible values for the 
parameters  and  additional filters of this function refer to 
https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- getMeta(object=cb, resource="species")
}

