% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getProteinInfo}
\alias{getProteinInfo}
\title{getProteinInfo}
\usage{
getProteinInfo(protein)
}
\arguments{
\item{protein}{a charcter vector of Uniprot Ids}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch annotations for specific protein/s
}
\examples{
res <- getProteinInfo("O15350")
}

