% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getSnpByGene}
\alias{getSnpByGene}
\title{getSnpByGene}
\usage{
getSnpByGene(gene)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch known variants (snps) for specific gene/s
}
\examples{
res <- getSnpByGene("TET1")
}

