% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getTfbsByRegion}
\alias{getTfbsByRegion}
\title{getTfbsByRegion}
\usage{
getTfbsByRegion(region)
}
\arguments{
\item{region}{a charcter vector of genomic regions, eg 17:1000000-1100000}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch Transcription facrots data for specific region/s
}
\examples{
res <- getTfbsByRegion("17:1000000-1189811")
}

