% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranscript-methods.R
\docType{methods}
\name{getTranscript,CellBaseR-method}
\alias{getTranscript,CellBaseR-method}
\alias{getTranscript}
\title{getTranscript}
\usage{
\S4method{getTranscript}{CellBaseR}(object, ids, resource, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the transcript ids to be queried, for example, ensemble transccript ID
like ENST00000380152}

\item{resource}{a character vector to specify the resource to be queried}

\item{param}{a object of class CellBaseParam specifying additional params for the query}
}
\value{
a dataframe with the results of the query
}
\description{
A method to query transcript data from Cellbase web services.
}
\details{
Please, for details on possible values for the 
parameters  and  additional params of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- getTranscript(object=cb, ids="ENST00000373644", resource="info")
}
