% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getTranscriptByGene}
\alias{getTranscriptByGene}
\title{getTranscriptByGene}
\usage{
getTranscriptByGene(gene)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch transcripts for specific gene/s
}
\examples{
res <- getTranscriptByGene("TET1")
}

