% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getVariantAnnotation}
\alias{getVariantAnnotation}
\title{getVariantAnnotation}
\usage{
getVariantAnnotation(variant)
}
\arguments{
\item{variant}{a charcter vector of length < 200 of genomic variants,
eg 19:45411941:T:C}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch variant annotation for specific variant/s
}
\examples{
res <- getVariantAnnotation("19:45411941:T:C")
}

