% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getCaddScores}
\alias{getCaddScores}
\title{getCaddScores}
\usage{
getCaddScores(variant)
}
\arguments{
\item{variant}{a charcter vector of genomic variants, eg 19:45411941:T:C}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch Cadd scores for specific variant/s
}
\examples{
res <- getCaddScores("19:45411941:T:C")
}

