% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getConservationByRegion}
\alias{getConservationByRegion}
\title{getConservationByRegion}
\usage{
getConservationByRegion(region, object = NULL, filters = NULL)
}
\arguments{
\item{region}{a charcter vector of genomic regions, eg 17:1000000-1100000}

\item{object}{an object of class CellBaseR}

\item{filters}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch conservation data for specific region/s
}
\examples{
res <- getConservationByRegion("17:1000000-1189811")
}

