% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getProteinInfo}
\alias{getProteinInfo}
\title{getProteinInfo}
\usage{
getProteinInfo(protein, object = NULL, filters = NULL)
}
\arguments{
\item{protein}{a charcter vector of Uniprot Ids}

\item{object}{an object of class CellBaseR}

\item{filters}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch annotations for specific protein/s
}
\examples{
res <- getProteinInfo("O15350")
}

