% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getSnpByGene}
\alias{getSnpByGene}
\title{getSnpByGene}
\usage{
getSnpByGene(gene, object = NULL, filters = NULL)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}

\item{object}{an object of class CellBaseR}

\item{filters}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch known variants (snps) for specific gene/s
}
\examples{
filters <- CellBaseParam(limit = 10)
res <- getSnpByGene("TET1", filters = filters)
}

