% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getTranscriptByGene}
\alias{getTranscriptByGene}
\title{getTranscriptByGene}
\usage{
getTranscriptByGene(gene, object = NULL, filters = NULL)
}
\arguments{
\item{gene}{a charcter vector of HUGO symbol (gene names)}

\item{object}{an object of class CellBaseR}

\item{filters}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch transcripts for specific gene/s
}
\examples{
res <- getTranscriptByGene("TET1")
}
